New TaskList class
