+++ src/main/java/Command/DisplayList.java
@@ -9,9 +9,9 @@ public class DisplayList extends Command {
 
   @Override
   void call_(Lmbd lmbd, String[] args) {
-    System.out.println(String.format("You have %d tasks:", lmbd.task_size()));
-    for (int i = 0; i < lmbd.task_size(); i++) {
-      System.out.println(String.format("%d. %s", i + 1, lmbd.get_task(i)));
+    System.out.println(String.format("You have %d tasks:", lmbd.TASKS.task_size()));
+    for (int i = 0; i < lmbd.TASKS.task_size(); i++) {
+      System.out.println(String.format("%d. %s", i + 1, lmbd.TASKS.get_task(i)));
     }
   }
 }
