+++ testdisk.log
@@ -1,0 +1,54 @@
+
+
+Wed Aug 20 20:03:07 2025
+Command line: TestDisk
+
+TestDisk 7.1, Data Recovery Utility, July 2019
+Christophe GRENIER <grenier@cgsecurity.org>
+https://www.cgsecurity.org
+OS: Linux, kernel 6.8.0-65-generic (#68~22.04.1-Ubuntu SMP PREEMPT_DYNAMIC Tue Jul 15 18:06:34 UTC 2) x86_64
+Compiler: GCC 11.2
+ext2fs lib: 1.46.5, ntfs lib: libntfs-3g, reiserfs lib: none, ewf lib: none, curses lib: ncurses 6.3
+User is not root!
+Hard disk list
+
+
+TestDisk exited normally.
+Using locale 'en_SG.UTF-8'.
+
+
+Wed Aug 20 20:03:08 2025
+Command line: TestDisk /debug
+
+TestDisk 7.1, Data Recovery Utility, July 2019
+Christophe GRENIER <grenier@cgsecurity.org>
+https://www.cgsecurity.org
+OS: Linux, kernel 6.8.0-65-generic (#68~22.04.1-Ubuntu SMP PREEMPT_DYNAMIC Tue Jul 15 18:06:34 UTC 2) x86_64
+Compiler: GCC 11.2
+ext2fs lib: 1.46.5, ntfs lib: libntfs-3g, reiserfs lib: none, ewf lib: none, curses lib: ncurses 6.3
+Warning: can't get size for Disk /dev/mapper/control - 0 B - 0 sectors, sector size=512
+Warning: can't get size for Disk /dev/loop18 - 0 B - 0 sectors, sector size=512
+Hard disk list
+Disk /dev/sda - 2000 GB / 1863 GiB - CHS 243201 255 63, sector size=512 - Realtek RTL9210 NVME, FW:1.00
+Disk /dev/loop0 - 4096 B - 8 sectors (RO), sector size=512
+Disk /dev/loop1 - 77 MB / 73 MiB - 151352 sectors (RO), sector size=512
+Disk /dev/loop10 - 304 MB / 290 MiB - 595512 sectors (RO), sector size=512
+Disk /dev/loop11 - 304 MB / 290 MiB - 595504 sectors (RO), sector size=512
+Disk /dev/loop12 - 13 MB / 12 MiB - 26472 sectors (RO), sector size=512
+Disk /dev/loop13 - 12 MB / 12 MiB - 24984 sectors (RO), sector size=512
+Disk /dev/loop14 - 53 MB / 50 MiB - 104240 sectors (RO), sector size=512
+Disk /dev/loop15 - 581 KB / 568 KiB - 1136 sectors (RO), sector size=512
+Disk /dev/loop16 - 51 MB / 49 MiB - 100952 sectors (RO), sector size=512
+Disk /dev/loop17 - 589 KB / 576 KiB - 1152 sectors (RO), sector size=512
+Disk /dev/loop2 - 77 MB / 73 MiB - 151376 sectors (RO), sector size=512
+Disk /dev/loop3 - 70 MB / 66 MiB - 136736 sectors (RO), sector size=512
+Disk /dev/loop4 - 70 MB / 66 MiB - 136776 sectors (RO), sector size=512
+Disk /dev/loop5 - 529 MB / 505 MiB - 1034424 sectors (RO), sector size=512
+Disk /dev/loop6 - 541 MB / 516 MiB - 1056784 sectors (RO), sector size=512
+Disk /dev/loop7 - 618 MB / 589 MiB - 1207552 sectors (RO), sector size=512
+Disk /dev/loop8 - 424 MB / 404 MiB - 828160 sectors (RO), sector size=512
+Disk /dev/loop9 - 96 MB / 91 MiB - 187776 sectors (RO), sector size=512
+Disk /dev/nvme0n1 - 1024 GB / 953 GiB - CHS 976762 64 32, sector size=512
+
+
+TestDisk exited normally.
